[#1793] Move existing FF circuits to ffa/ submodule
