+++ msm/src/ffa/columns.rs
@@ -1,0 +1,28 @@
+use crate::columns::Column;
+use crate::columns::ColumnIndexer;
+
+use crate::LIMBS_NUM;
+
+#[derive(Clone, Copy, Debug, PartialEq)]
+/// Column indexer for MSM columns
+pub enum MSMColumnIndexer {
+    A(usize),
+    B(usize),
+    C(usize),
+    D(usize),
+}
+
+impl ColumnIndexer for MSMColumnIndexer {
+    fn ix_to_column(self) -> Column {
+        let to_column_inner = |offset, i| {
+            assert!(i < LIMBS_NUM);
+            Column::X(LIMBS_NUM * offset + i)
+        };
+        match self {
+            MSMColumnIndexer::A(i) => to_column_inner(0, i),
+            MSMColumnIndexer::B(i) => to_column_inner(1, i),
+            MSMColumnIndexer::C(i) => to_column_inner(2, i),
+            MSMColumnIndexer::D(i) => to_column_inner(3, i),
+        }
+    }
+}
